Lie Algebras API

Creating Algebras

Kiwi.A_series — Method

A_series(n::Int)

Construct the A_n Lie algebra (su(n+1)).

Kiwi.B_series — Method

B_series(n::Int)

Construct the B_n Lie algebra (so(2n+1)).

Kiwi.C_series — Method

C_series(n::Int)

Construct the C_n Lie algebra (sp(2n)).

Kiwi.D_series — Method

D_series(n::Int)

Construct the D_n Lie algebra (so(2n)).

Kiwi.E_series — Method

E_series(n::Int)

Construct the E_n Lie algebra (n = 6, 7, or 8).

Kiwi.F_series — Method

F_series(n::Int)

Construct the F_4 Lie algebra (n must be 4).

Kiwi.G_series — Method

G_series(n::Int)

Construct the G_2 Lie algebra (n must be 2).

Algebra Properties

Kiwi.cartan_matrix — Method

cartan_matrix(g::LieAlgebra)

Return the Cartan matrix for the given Lie algebra.

Kiwi.dual_coxeter_number — Method

dual_coxeter_number(g::LieAlgebra)

Compute the dual Coxeter number hⱽ of the Lie algebra. This is defined as the Dynkin index of the adjoint representation.

Kiwi.lie_rank — Method

lie_rank(g::LieAlgebra)

Return the rank of the Lie algebra.

Root System

Kiwi.simple_roots — Method

simple_roots(g::LieAlgebra)

Return the simple roots of the Lie algebra in the fundamental weight basis. The simple root αi in fundamental weight coordinates: αi = Σj Cij ω_j (rows of C)