Kiwi

Documentation for Kiwi.

Kiwi.add_constraint

Kiwi.add_edit_variable

Kiwi.add_with_artificial_variable

Kiwi.all_dummies

Kiwi.any_pivotable_symbol

Kiwi.choose_subject

Kiwi.coefficient

Kiwi.create_row

Kiwi.dual_optimize

Kiwi.get_dual_entering_symbol

Kiwi.get_entering_symbol

Kiwi.get_leaving_row

Kiwi.get_marker_leaving_row

Kiwi.hasConstraint

Kiwi.has_edit_variable

Kiwi.insert!

Kiwi.insert!

Kiwi.optimize

Kiwi.remove!

Kiwi.remove_constraint

Kiwi.remove_constraint_effects

Kiwi.remove_edit_variable

Kiwi.remove_marker_effects

Kiwi.reverse_sign!

Kiwi.runtests

Kiwi.solve!

Kiwi.solve!

Kiwi.substitute!

Kiwi.suggest_value

Kiwi.update_variables

Kiwi.add_constraint — Method

Add a constraint to the solver.

Errors:

DuplicateConstraint: The given constraint has already been added to the solver.

UnsatisfiableConstraint: The given constraint is required and cannot be satisfied.

Kiwi.add_edit_variable — Method

Add an edit variable to the solver.

This method should be called before the suggestValue method is used to supply a suggested value for the given edit variable.

Errors:

DuplicateEditVariable: The given edit variable has already been added to the solver.

BadRequiredStrength: The given strength is >= required.

Kiwi.add_with_artificial_variable — Method

Add the row to the tableau using an artificial variable.

This will return false if the constraint cannot be satisfied.

Kiwi.all_dummies — Method

Test whether a row is composed of all dummy variables.

Kiwi.any_pivotable_symbol — Method

Get the first Slack or Error symbol in the row.

If no such symbol is present, and Invalid symbol will be returned.

Kiwi.choose_subject — Method

Choose the subject for solving for the row.

This method will choose the best subject for using as the solve target for the row. An invalid symbol will be returned if there is no valid target.

The symbols are chosen according to the following precedence:

The first symbol representing an external variable.

A negative slack or error tag variable.

If a subject cannot be found, an invalid symbol will be returned.

Kiwi.coefficient — Method

Get the coefficient for the given symbol.

If the symbol does not exist in the row, zero will be returned.

Kiwi.create_row — Method

Create a new Row object for the given constraint.

The terms in the constraint will be converted to cells in the row. Any term in the constraint with a coefficient of zero is ignored. This method uses the getVarSymbol method to get the symbol for the variables added to the row. If the symbol for a given cell variable is basic, the cell variable will be substituted with the basic row.

The necessary slack and error variables will be added to the row. If the constant for the row is negative, the sign for the row will be inverted so the constant becomes positive.

The tag will be updated with the marker and error symbols to use for tracking the movement of the constraint in the tableau.

Kiwi.dual_optimize — Method

Optimize the system using the dual of the simplex method.

The current state of the system should be such that the objective function is optimal, but not feasible. This method will perform an iteration of the dual simplex method to make the solution both optimal and feasible.

Errors:

InternalSolverError: The system cannot be dual optimized.

Kiwi.get_dual_entering_symbol — Method

Compute the entering symbol for the dual optimize operation.

This method will return the symbol in the row which has a positive coefficient and yields the minimum ratio for its respective symbol in the objective function. The provided row must be infeasible. If no symbol is found which meats the criteria, an invalid symbol is returned.

Kiwi.get_entering_symbol — Method

Compute the entering variable for a pivot operation.

This method will return first symbol in the objective function which is non-dummy and has a coefficient less than zero. If no symbol meets the criteria, it means the objective function is at a minimum, and an invalid symbol is returned.

Kiwi.get_leaving_row — Method

Compute the row which holds the exit symbol for a pivot.

This method will return an iterator to the row in the row map which holds the exit symbol. If no appropriate exit symbol is found, the end() iterator will be returned. This indicates that the objective function is unbounded.

Kiwi.get_marker_leaving_row — Method

Compute the leaving row for a marker variable.

This method will return an iterator to the row in the row map which holds the given marker variable. The row will be chosen according to the following precedence:

The row with a restricted basic varible and a negative coefficient for the marker with the smallest ratio of -constant / coefficient.

The row with a restricted basic variable and the smallest ratio of constant / coefficient.

The last unrestricted row which contains the marker.

If the marker does not exist in any row, the row map end() iterator will be returned. This indicates an internal solver error since the marker should exist somewhere in the tableau.

Kiwi.hasConstraint — Method

Test whether a constraint has been added to the solver.

Kiwi.has_edit_variable — Method

Test whether an edit variable has been added to the solver.

Kiwi.insert! — Function

Insert a symbol into the row with a given coefficient.

If the symbol already exists in the row, the coefficient will be added to the existing coefficient. If the resulting coefficient is zero, the symbol will be removed from the row.

Kiwi.insert! — Function

Insert a row into this row with a given coefficient.

The constant and the cells of the other row will be multiplied by the coefficient and added to this row. Any cell with a resulting coefficient of zero will be removed from the row.

Kiwi.optimize — Method

Optimize the system for the given objective function.

This method performs iterations of Phase 2 of the simplex method until the objective function reaches a minimum.

Errors:

InternalSolverError: The value of the objective function is unbounded.

Kiwi.remove! — Method

Remove the given symbol from the row.

Kiwi.remove_constraint — Method

Remove a constraint from the solver.

Errors:

UnknownConstraint: The given constraint has not been added to the solver.

Kiwi.remove_constraint_effects — Method

Remove the effects of a constraint on the objective function.

Kiwi.remove_edit_variable — Method

Remove an edit variable from the solver.

Error:

UnknownEditVariable: The given edit variable has not been added to the solver.

Kiwi.remove_marker_effects — Method

Remove the effects of an error marker on the objective function.

Kiwi.reverse_sign! — Method

Reverse the sign of the constant and all cells in the row.

Kiwi.runtests — Method

Kiwi.runtests(pattern...; kwargs...)

Equivalent to ReTest.retest(Kiwi, pattern...; kwargs...). This function is defined automatically in any module containing a @testset, possibly nested within submodules.

Kiwi.solve! — Method

Solve the row for the given symbols.

This method assumes the row is of the form x = b * y + c and will solve the row such that y = x / b - c / b. The rhs symbol will be removed from the row, the lhs added, and the result divided by the negative inverse of the rhs coefficient.

The lhs symbol must not exist in the row, and the rhs symbol must exist in the row.

Kiwi.solve! — Method

Solve the row for the given symbol.

This method assumes the row is of the form a * x + b * y + c = 0 and (assuming solve for x) will modify the row to represent the right hand side of x = -b/a * y - c / a. The target symbol will be removed from the row, and the constant and other cells will be multiplied by the negative inverse of the target coefficient.

The given symbol must exist in the row.

Kiwi.substitute! — Method

Substitute the parametric symbol with the given row.

This method will substitute all instances of the parametric symbol in the tableau and the objective function with the given row.

Kiwi.suggest_value — Method

Suggest a value for the given edit variable.

This method should be used after an edit variable as been added to the solver in order to suggest the value for that variable.

Errors:

UnknownEditVariable: The given edit variable has not been added to the solver.

Kiwi.update_variables — Method

Update the values of the external solver variables.